backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2
requests==2.31.0
openai==1.3.0
chromadb==0.4.15
numpy==1.24.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4